headers/os: Make headers generator-friendly
Make Haiku headers a bit more friendly to binding generators by:
- Giving some `enum`s names (especially those that appear in
default arguments).
- Converting some closely-related constant integers into enums.
- Converting an internal `inline` function into a macro so that
the result could be evaluated in compile time.

+++ src/kits/tracker/FindPanel.cpp
@@ -2826,7 +2826,7 @@ FindPanel::AddLogicMenu(int32 index, bool selectAnd)
 
 	menu->SetTargetForItems(this);
 
-	BMenuField* menufield = new BMenuField("Logic", "", menu, B_WILL_DRAW);
+	BMenuField* menufield = new BMenuField("Logic", "", menu, (uint32)B_WILL_DRAW);
 	menufield->SetDivider(0.0f);
 
 	ResizeMenuField(menufield);
